+++  course-work-ps-31-2013 --username dertkaes@gmail.com/OpenGLTest/OpenGLTest/view/Camera.h
@@ -1,28 +1,28 @@
 class CCamera
 {
 public:
-	static const double SCALE;
-	static const double TRANSLATE;
+	static const float SCALE;
+	static const float TRANSLATE;
 
 	CCamera(void):m_maxTransX(8.0), m_maxTransY(8.0), m_maxScale(2.8), m_minScale(0.35) { Reset(); }
 	void Update();
-	void Translate(double transX, double transY);
-	void Rotate(double rotZ, double rotX);
+	void Translate(float transX, float transY);
+	void Rotate(float rotZ, float rotX);
 	void ZoomIn();
 	void ZoomOut();
 	void Reset();
-	void SetLimits(double maxTransX, double maxTransY, double maxScale, double minScale);
-	double GetTranslationX() { return m_transX; }
-	double GetTranslationY() { return m_transY; }
-	double GetScale() { return m_scale; }
+	void SetLimits(float maxTransX, float maxTransY, float maxScale, float minScale);
+	float GetTranslationX() { return m_transX; }
+	float GetTranslationY() { return m_transY; }
+	float GetScale() { return m_scale; }
 private:
-	double m_maxTransX;
-	double m_maxTransY;
-	double m_maxScale;
-	double m_minScale;
-	double m_transX;
-	double m_transY;
-	double m_rotX;
-	double m_rotZ;
-	double m_scale;
+	float m_maxTransX;
+	float m_maxTransY;
+	float m_maxScale;
+	float m_minScale;
+	float m_transX;
+	float m_transY;
+	float m_rotX;
+	float m_rotZ;
+	float m_scale;
 };
