+++  course-work-ps-31-2013 --username dertkaes@gmail.com/OpenGLTest/OpenGLTest/view/OBJModelFactory.cpp
@@ -178,7 +178,7 @@ C3DModel * LoadObjModel(std::string const& path)
 		newNormals.swap(normals);
 	}
 	std::string boundingPath = path.substr(0, path.find_last_of('.')) + ".txt";
-	double scale = 1.0;
+	float scale = 1.0;
 	std::shared_ptr<IBounding> bounding = LoadBoundingFromFile(boundingPath, scale);
 	return new C3DModel(newVertices, newTextureCoords, newNormals, indexes, materialManager, meshes, bounding, scale);
 }
