+++  course-work-ps-31-2013 --username dertkaes@gmail.com/OpenGLTest/OpenGLTest/view/Vector2.h
@@ -192,4 +192,4 @@ CVector2<T> const operator*(T const& scalar, CVector2<T> const& v) throw()
 }
 
 typedef CVector2<float> CVector2f;
-typedef CVector2<double> CVector2d;
+typedef CVector2<float> CVector2d;
